test: all legacy test
